Rename to match model type.
