+++ mlx_audio/stt/models/__init__.py
@@ -3,7 +3,7 @@ from . import (
     fireredasr2,
     glmasr,
     granite_speech,
-    granite_speech5,
+    granite_speech5_ctc,
     granite_speech_nar,
     lasr_ctc,
     moss_music,
