+++ mlx_audio/stt/models/granite_speech5_ctc/config.py
