+++ mlx_audio/stt/models/granite_speech5_ctc/granite_speech5.py
