+++ mlx_audio/stt/utils.py
@@ -95,7 +95,7 @@ MODEL_REMAPPING = {
     "moonshine": "moonshine",
     "mms": "mms",
     "granite_speech": "granite_speech",
-    "granite_speech5_ctc": "granite_speech5",
+    "granite_speech5_ctc": "granite_speech5_ctc",
     "granite_speech_nar": "granite_speech_nar",
     "qwen2_audio": "qwen2_audio",
     "mega_asr": "mega_asr",
